requirements.txt
Flask==2.3.2
google-cloud-bigquery==3.12.1
google-auth==2.21.0
gunicorn==20.1.0

